requirements.txt
streamlit
torch==2.1.0
torchvision==0.16.0
pillow
scikit-learn
